requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.4
django-cors-headers==3.12.0
django-environ==0.8.1
django-oauth-toolkit==2.0.0
djangorestframework==3.13.1
drf-social-oauth2==1.0.8
idna==3.3
jwcrypto==1.3
oauthlib==3.2.0
pycparser==2.21
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
urllib3==1.26.9
wrapt==1.14.1
